Workaround for slices disappearing
I'm pretty sure this bug wasn't my fault and is an issue with JavaFX. If
you animate to the location that it already is, i.e., if MoveTo and
LineTo are to the same location, the animated Node disappears. Weird.
So, to fix that, I just make sure they're never the same.